+++ CHANGELOG.md
@@ -5,5 +5,7 @@
 
 ## Features
  * support for monitoring ehcache (stagemonitor-ehcache)
+ * overwrite properties with Java system property (-Dstagemonitor.property=value)
+ * disable plugins with property 'stagemonitor.plugins.disabled'
 
 ## Bug Fixes